+++ src/IOEither.ts
@@ -10,7 +10,7 @@ import { Apply2 } from './Apply'
 import { Bifunctor2 } from './Bifunctor'
 import { Compactable2C, getCompactableComposition } from './Compactable'
 import * as E from './Either'
-import { Filterable2C, getFilterableComposition } from './Filterable'
+import { Filterable2C } from './Filterable'
 import { bindTo_, bind_, flow, identity, Lazy, pipe, Predicate, Refinement, tuple } from './function'
 import { Functor2 } from './Functor'
 import * as I from './IO'
@@ -18,7 +18,7 @@ import { Monad2 } from './Monad'
 import { MonadIO2 } from './MonadIO'
 import { MonadThrow2 } from './MonadThrow'
 import { Monoid } from './Monoid'
-import { Option } from './Option'
+import { getLeft, getRight, Option } from './Option'
 import { Semigroup } from './Semigroup'
 
 // -------------------------------------------------------------------------------------
@@ -511,14 +511,33 @@ export function getCompactable<E>(M: Monoid<E>): Compactable2C<URI, E> {
  * @since 2.1.0
  */
 export function getFilterable<E>(M: Monoid<E>): Filterable2C<URI, E> {
-  const F = getFilterableComposition(I.Functor, E.getFilterable(M))
+  const F = E.getFilterable(M)
+
+  const filter = <A>(fa: IOEither<E, A>, predicate: Predicate<A>): IOEither<E, A> =>
+    pipe(
+      fa,
+      I.map((ga) => F.filter(ga, predicate))
+    )
+  const filterMap = <A, B>(fa: IOEither<E, A>, f: (a: A) => Option<B>) =>
+    pipe(
+      fa,
+      I.map((ga) => F.filterMap(ga, f))
+    )
 
   return {
     URI,
-    filter: F.filter,
-    filterMap: F.filterMap,
-    partition: F.partition,
-    partitionMap: F.partitionMap
+    filter,
+    filterMap,
+    partition: <A>(fa: IOEither<E, A>, predicate: Predicate<A>) => {
+      const left = filter(fa, (a) => !predicate(a))
+      const right = filter(fa, predicate)
+      return { left, right }
+    },
+    partitionMap: (fa, f) => {
+      const left = filterMap(fa, (a) => getLeft(f(a)))
+      const right = filterMap(fa, (a) => getRight(f(a)))
+      return { left, right }
+    }
   }
 }